Landable::ApiResponder
Enabling the use of `respond_with` in combination with namespace support
when looking up the serializer to be used.

This could have been defined with a `Page.active_model_serializer` class
method on each model, but I have Bigger Plans(tm) that make a custom
responder a more reasonable option.